requirements.txt
brotlipy==0.7.0
certifi @ file:///opt/conda/conda-bld/certifi_1663615672595/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cryptography @ file:///tmp/build/80754af9/cryptography_1652101588893/work
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
psycopg2-binary==2.9.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.2
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
typing_extensions==4.3.0
urllib3 @ file:///tmp/abs_5dhwnz6atv/croots/recipe/urllib3_1659110457909/work
vk-api==11.9.9
